# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Spacca POS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stock/movements:
    get:
      operationId: listStockMovements
      tags: [stock]
      summary: List stock movements ledger
      parameters:
        - in: query
          name: ingredientId
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        "200":
          description: Stock movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockMovement"
  /stock/adjustments:
    post:
      operationId: createStockAdjustment
      tags: [stock]
      summary: Create manual stock adjustment or waste entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockAdjustmentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockMovement"
...
  /users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDetail"
    post:
      operationId: createUser
      tags: [admin]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetail"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [admin]
      summary: Update a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: query
          name: branchId
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetail"
    delete:
      operationId: deleteUser
      tags: [admin]
      summary: Delete a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: query
          name: branchId
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    UserDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        role:
          type: string
        branchId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - username
        - role
        - isActive
...
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, barista, frontdesk, cashier, pickup, stockcontrol]
        branchId:
          type: integer
        pin:
          type: string
      required:
        - name
        - username
        - password
        - role
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, barista, frontdesk, cashier, pickup]
        pin:
          type: string
        isActive:
          type: boolean
...
    StockMovement:
      type: object
      properties:
        id:
          type: integer
        ingredientId:
          type: integer
        ingredientName:
          type: string
        orderId:
          type: ["integer", "null"]
        movementType:
          type: string
          enum: [sale, restock, adjustment, waste, opening]
        quantity:
          type: number
        quantityAfter:
          type: number
        note:
          type: ["string", "null"]
        createdBy:
          type: integer
        createdByName:
          type: string
        createdAt:
          type: string
      required:
        - id
        - ingredientId
        - ingredientName
        - orderId
        - movementType
        - quantity
        - quantityAfter
        - note
        - createdBy
        - createdByName
        - createdAt
...
    StockAdjustmentBody:
      type: object
      properties:
        ingredientId:
          type: integer
        movementType:
          type: string
          enum: [adjustment, waste, opening]
        quantity:
          type: number
        note:
          type: string
      required:
        - ingredientId
        - movementType
        - quantity
...